+++ src/main/java/server/preonboarding/budgetmanager/security/JwtProvider.java
@@ -1,27 +1,15 @@
 package server.preonboarding.budgetmanager.security;
 
-import io.jsonwebtoken.*;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
+import io.jsonwebtoken.Jwts;
+import lombok.RequiredArgsConstructor;
 
 import javax.crypto.SecretKey;
 import java.util.Date;
 
-@Slf4j
-@Component
+@RequiredArgsConstructor
 public class JwtProvider {
 
     private final SecretKey secretKey;
-    private final JwtParser jwtParser;
-
-    public JwtProvider(@Value("${jwt.secret}") String secret) {
-        byte[] bytes = Decoders.BASE64.decode(secret);
-        secretKey = Keys.hmacShaKeyFor(bytes);
-        jwtParser = Jwts.parser().verifyWith(secretKey).build();
-    }
 
     public String createAccessToken(Long id) {
         Date now = new Date();
@@ -30,17 +18,6 @@ public class JwtProvider {
                 .issuedAt(now)
                 .signWith(secretKey)
                 .compact();
-    }
-
-    public Long extractSubject(String accessToken) {
-        String id = getJwsClaims(accessToken)
-                .getPayload()
-                .getSubject();
-        return Long.parseLong(id);
-    }
-
-    private Jws<Claims> getJwsClaims(String accessToken) {
-        return jwtParser.parseSignedClaims(accessToken);
     }
 
 }
